backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlmodel==0.0.14
passlib[bcrypt]==1.7.4
